requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
celery==5.2.7
python-dotenv==0.20.0
Flask-SQLAlchemy==3.0.3
Flask-DB==0.3.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
passlib==1.7.4
itsdangerous==2.1.2
marshmallow-sqlalchemy==0.29.0
Flask-Mail==0.9.1
Flask-JWT-Extended==4.4.4
flask-redis==0.4.0
requests==2.28.2
numpy==1.24.2